Rename legacy OnboardingViewController
